[NFC] Expose internal LLVMGPU APIs for vector_distribute
This allows external systems to make use of IREE's vector_distribute functionality in a composable manner.
This depends on https://github.com/llvm/llvm-project/pull/145313 landing and being integrated into IREE.
