+++ compiler/src/iree/compiler/Codegen/LLVMGPU/TransformExtensions/LLVMGPUExtensions.cpp
@@ -1458,16 +1458,6 @@ transform_dialect::PrefetchSharedMemoryCopiesOp::applyToOne(
   return DiagnosedSilenceableFailure::success();
 }
 
-class TransformVectorLayoutOptions : public VectorLayoutOptions {
-public:
-  TransformVectorLayoutOptions(Operation *root, bool fullConversion)
-      : VectorLayoutOptions(root, fullConversion) {}
-
-  VectorLayoutInterface getDefaultLayout(VectorType type) const override {
-    return VectorLayoutInterface();
-  }
-};
-
 DiagnosedSilenceableFailure
 transform_dialect::AMDGPUDistributeVectorsOp::applyToOne(
     transform::TransformRewriter &rewriter, mlir::FunctionOpInterface target,
