+++ compiler/src/iree/compiler/Codegen/LLVMGPU/Utils/LLVMGPUUtils.h
@@ -7,11 +7,34 @@
 #ifndef IREE_COMPILER_CODEGEN_LLVMGPU_UTILS_LLVMGPUUTILS_H_
 #define IREE_COMPILER_CODEGEN_LLVMGPU_UTILS_LLVMGPUUTILS_H_
 
+#include "iree/compiler/Codegen/Common/GPU/GPUVectorDistribution.h"
 #include "mlir/Dialect/SCF/IR/SCF.h"
 #include "mlir/IR/PatternMatch.h"
 #include "mlir/Interfaces/FunctionInterfaces.h"
 
-namespace mlir::iree_compiler {
+namespace mlir {
+
+namespace linalg {
+class LinalgOp;
+} // namespace linalg
+
+namespace vector {
+class ContractionOp;
+} // namespace vector
+
+namespace iree_compiler {
+class VectorContractOpInfo;
+
+class ContractionVectorLayoutOptions : public VectorLayoutOptions {
+public:
+  ContractionVectorLayoutOptions(Operation *root, Value laneId,
+                                 int64_t subgroupSize);
+  RewritePatternSet &getPatterns();
+  VectorLayoutInterface getDefaultLayout(VectorType type) const override;
+
+private:
+  RewritePatternSet patterns;
+};
 
 /// Helper to convert copy to shared memory to async copy. This creates groups
 /// of consecutive copies and emit wait operation right after.
@@ -39,6 +62,25 @@ FailureOr<scf::ForOp> prefetchSharedMemoryCopy(RewriterBase &rewriter,
 void addBarrier(mlir::FunctionOpInterface funcOp, Operation *alloc,
                 ArrayRef<Operation *> aliasGroup, bool hasAsyncCopies = true);
 
-} // namespace mlir::iree_compiler
+namespace IREE {
+namespace GPU {
+class MMAScheduleAttr;
+
+::llvm::FailureOr<::std::tuple<IREE::VectorExt::VectorLayoutInterface,
+                               IREE::VectorExt::VectorLayoutInterface,
+                               IREE::VectorExt::VectorLayoutInterface>>
+getContractionLayout(IREE::GPU::MMAScheduleAttr scheduleAttr,
+                     VectorContractOpInfo &opInfo, linalg::LinalgOp contractOp);
+
+::llvm::FailureOr<::std::tuple<IREE::VectorExt::VectorLayoutInterface,
+                               IREE::VectorExt::VectorLayoutInterface,
+                               IREE::VectorExt::VectorLayoutInterface>>
+getContractionLayout(IREE::GPU::MMAScheduleAttr scheduleAttr,
+                     VectorContractOpInfo &opInfo,
+                     vector::ContractionOp contractOp);
+} // namespace GPU
+} // namespace IREE
+} // namespace iree_compiler
+} // namespace mlir
 
 #endif
