+++ compiler/src/iree/compiler/Utils/BUILD.bazel
@@ -58,9 +58,7 @@ iree_compiler_cc_library(
         "//runtime/src/iree/base/internal/flatcc:parsing",
         "@llvm-project//llvm:Passes",
         "@llvm-project//llvm:Support",
-        "@llvm-project//mlir:AffineDialect",
         "@llvm-project//mlir:ArithDialect",
-        "@llvm-project//mlir:DialectUtils",
         "@llvm-project//mlir:FuncDialect",
         "@llvm-project//mlir:FunctionInterfaces",
         "@llvm-project//mlir:IR",
