+++ compiler/src/iree/compiler/Utils/CMakeLists.txt
@@ -46,7 +46,6 @@ iree_cc_library(
   DEPS
     LLVMPasses
     LLVMSupport
-    MLIRAffineDialect
     MLIRArithDialect
     MLIRFuncDialect
     MLIRFunctionInterfaces
